requirements-dev.txt
Babel==2.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
djangorestframework==3.9.3
python-dateutil==2.8.0
django-rest-swagger==2.2.0
